requirements.txt
sqlalchemy-iris
langchain
langchain-iris
langchain-openai
streamlit
streamlit-chat
openai
pypdf
